rename package to mrconcurrent
