--- 6.824/go.mod
@@ -1,3 +1,0 @@
-module github.com/mit-distributed-systems/6.824
-
-go 1.16
