+++ 6.824/src/main/mrworker.go
@@ -15,7 +15,7 @@ import (
 	"plugin"
 
 	logger "github.com/mit-distributed-systems/6.824/src/logger"
-	mr "github.com/mit-distributed-systems/6.824/src/mr"
+	mr "github.com/mit-distributed-systems/6.824/src/mrconcurrent"
 	log "github.com/sirupsen/logrus"
 )
