+++ 6.824/src/mrconcurrent/rpc.go
@@ -1,4 +1,4 @@
-package mr
+package mrconcurrent
 
 //
 // RPC definitions.
